+++ Domain/Models/TodayOdr/ITodayOdrRepository.cs
@@ -26,7 +26,7 @@ public interface ITodayOdrRepository : IRepositoryBase
 
     Dictionary<string, string> CheckNameChanged(List<OrdInfModel> odrInfModelList);
 
-    List<(int type, string itemName, int lastDaySanteiRiha, string rihaItemName)> GetValidGairaiRiha(int hpId, int ptId, long raiinNo, int sinDate, int syosaiKbn, List<Tuple<string, string>> allOdrInfItems);
+    List<(int type, string itemName, int lastDaySanteiRiha, string rihaItemName)> GetValidGairaiRiha(int hpId, long ptId, long raiinNo, int sinDate, int syosaiKbn, List<Tuple<string, string>> allOdrInfItems);
 
     (double systemSetting, bool isExistYoboItemOnly) GetValidJihiYobo(int hpId, int syosaiKbn, int sinDate, List<string> itemCds);
