+++ EmrCloudApi/Requests/Document/CheckExistFileNameRequest.cs
@@ -6,7 +6,7 @@ public class CheckExistFileNameRequest
 
     public int CategoryCd { get; set; }
 
-    public int PtId { get; set; }
+    public long PtId { get; set; }
 
     public bool IsCheckDocInf { get; set; }
 }
