+++ EmrCloudApi/Requests/VisitingList/GetReceptionLockRequest.cs
@@ -2,7 +2,7 @@
 {
     public class GetReceptionLockRequest
     {
-        public long SinDate { get; set; }
+        public int SinDate { get; set; }
         public long PtId { get; set; }
         public long RaiinNo { get; set; }
         public string FunctionCd { get; set; } = string.Empty;
